requirements.txt
todoist-python==8.1.3
python-dotenv==0.19.0
click==8.0.1
SQLAlchemy==1.4.22
faker==8.11.0
